requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
openai==1.6.1
line-bot-sdk==3.5.1
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
aiosqlite==0.19.0
httpx==0.25.2
jinja2==3.1.2
psycopg2-binary==2.9.9
gunicorn==21.2.0
